RAG_RESEARCH_AGENT/requirements-dev.txt
-r requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
moto[all]>=5.0.0
boto3-stubs[essential]>=1.35.0
bandit>=1.7.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
